Reorganize model1_maskrcnn scripts.
